Refactor: move fromReconcileStore to its own module
